fix e2e increase resolution
