+++ lib/testing/src/lib/process-services-cloud/pages/edit-process-filter-cloud-component.page.ts
@@ -78,7 +78,7 @@ export class EditProcessFilterCloudComponentPage {
 
         const orderElement = element.all(by.cssContainingText('mat-option span', option)).first();
         await BrowserActions.click(orderElement);
-        await browser.sleep(1000);
+        await browser.sleep(1500);
     }
 
     async getOrderFilterDropDownValue(): Promise<string> {
